pyproject.toml
[project]
name = "ydc-ai-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agentscope>=1.0.5",
    "autogen-ext>=0.7.5",
    "langchain>=1.0.1",
    "langchain-experimental>=0.0.42",
    "langchain-openai>=1.0.0",
    "langgraph>=1.0.0",
    "pyautogen>=0.10.0",
    "python-dotenv>=1.1.1",
    "mcp>=1.0.0",
    "pytz>=2024.1",
    "langchain-mcp-adapters>=0.1.0",
    "requests>=2.31.0",
    "pydantic>=2.12.3",
    "langchain-deepseek>=1.0.0",
    "ruff>=0.14.4",
]
